requirements.txt
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
itsdangerous==2.2.0
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.6.0
WTForms==3.2.1
email_validator==2.2.0
